requirements.txt
requests
beautifulsoup4
mecab-python3
ipadic
wordcloud
tqdm